+ make genome complexity a float value + serialization updated + reset age after cell activation

+++ source/EngineGpuKernels/AttackerProcessor.cuh
@@ -84,7 +84,7 @@ __device__ __inline__ void AttackerProcessor::processCell(SimulationData& data, 
                     color,
                     otherColor);
                 energyToTransfer /=
-                    (1.0f + cellFunctionAttackerGenomeComplexityBonus * static_cast<float>(otherCell->genomeComplexity - cell->genomeComplexity));
+                    (1.0f + cellFunctionAttackerGenomeComplexityBonus * (otherCell->genomeComplexity - cell->genomeComplexity));
             }
             if (cudaSimulationParameters.features.advancedAttackerControl
                 && ((otherCell->mutationId == cell->mutationId) || (otherCell->ancestorMutationId == static_cast<uint8_t>(cell->mutationId & 0xff)))
