+++ source/EngineGpuKernels/TOs.cuh
@@ -164,7 +164,7 @@ struct CellTO
     uint32_t creatureId;
     uint32_t mutationId;
     uint8_t ancestorMutationId;  //only the first 8 bits from ancestor mutation id
-    uint32_t genomeComplexity;
+    float genomeComplexity;
 
     //cell function
     uint8_t executionOrderNumber;
